+++ include/solvers/navier_stokes_scratch_data.h
@@ -197,7 +197,8 @@ public:
         sd.fe_values.get_mapping(),
         sd.fe_values_particle_drag->get_fe(),
         sd.fe_values_particle_two_way_coupling_force->get_fe(),
-        sd.fe_values_particle_velocity->get_fe());
+        sd.fe_values_particle_velocity->get_fe(),
+        sd.fe_values_particle_momentum_transfer_coefficient->get_fe());
 
     gather_hessian = sd.gather_hessian;
   }
@@ -600,19 +601,24 @@ public:
                        const Mapping<dim>       &mapping);
 
   /**
-   * @brief enable_particle_field_projection Enables the collection of the particle fields
-   * projection data by the scratch
+   * @brief enable_particle_field_projection Enables the collection of the particle
+   * fields projection data by the scratch
    *
-   * @param quadrature Quadrature rule of the Navier-Stokes problem assembly
+   * @param[in] quadrature Quadrature rule of the Navier-Stokes problem assembly
    *
-   * @param mapping Mapping used for the Navier-Stokes problem assembly
+   * @param[in] mapping Mapping used for the Navier-Stokes problem assembly
    *
-   * @param fe_particle_drag_proj FiniteElement associated with the projected particle drag force
+   * @param[in] fe_particle_drag_proj FiniteElement associated with the
+   * projected particle drag force
    *
-   * @param fe_particle_two_way_coupling_force_proj FiniteElement associated with the projected
-   * particle two-way coupling force
+   * @param[in] fe_particle_two_way_coupling_force_proj FiniteElement associated
+   * with the projected particle two-way coupling force
    *
-   * @param fe_particle_velocity_proj FiniteElement associated with the projected particle velocity
+   * @param[in] fe_particle_velocity_proj FiniteElement associated with the
+   * projected particle velocity
+   *
+   * @param[in] fe_particle_momentum_transfer_coefficient FiniteElement
+   * associated with the projected particle momentum transfer coefficient
    */
   void
   enable_particle_field_projection(
@@ -620,7 +626,8 @@ public:
     const Mapping<dim>       &mapping,
     const FiniteElement<dim> &fe_particle_drag_proj,
     const FiniteElement<dim> &fe_particle_two_way_coupling_force_proj,
-    const FiniteElement<dim> &fe_particle_velocity_proj);
+    const FiniteElement<dim> &fe_particle_velocity_proj,
+    const FiniteElement<dim> &fe_particle_momentum_transfer_coefficient);
 
   /**
    *  @brief Reinitialize the content of the scratch for the void fraction
@@ -1314,21 +1321,33 @@ public:
    * @param[in] particle_drag_cell Iterator pointing to the current active cell
    * using the particle drag force DoFHandler
    *
-   * @param[in] particle_two_way_coupling_force Iterator pointing to the current
-   * active cell using the particle two-way coupling force DoFHandler
+   * @param[in] particle_two_way_coupling_force_cell Iterator pointing to the
+   * current active cell using the particle two-way coupling force DoFHandler
    *
    * @param[in] particle_velocity_cell Iterator pointing to the current active
    * cell using the particle velocity DoFHandler
    *
-   * @param[in] particle_fluid_drag Object containing the projection of the
-   * particle drag onto the fluid dofs
+   * @param[in] particle_momentum_transfer_coefficient_cell Iterator pointing to
+   * the current active cell using the particle momentum transfer coefficient
+   * DoFHandler
    *
-   * @param[in] particle_two_way_coupling_force Object containing the projection
-   * of the two-way coupling force projected from the particles onto the fluid
-   * dofs
+   * @param[in] particle_fluid_drag Object containing the projection of the drag
+   * calculated for the particles onto the fluid dofs
+   *
+   * @param[in] particle_fluid_force_two_way_coupling Object containing the
+   * projection of the two-way coupling force calculated for the particles onto
+   * the fluid dofs
    *
    * @param[in] particle_velocity Object containing the projection of the
    * particle velocities onto the fluid dofs
+   *
+   * @param[in] particle_momentum_transfer_coefficient Object containing the
+   * projection of the momentum transfer coefficient calculated for the
+   * particles onto the fluid dofs
+   *
+   * @param[in] drag_coupling Enumeration specifying the numerical coupling
+   * strategy used for the computation of the drag force between the fluid and
+   * the particles.
    */
   template <typename VectorType>
   void
@@ -1337,10 +1356,13 @@ public:
     const typename DoFHandler<dim>::active_cell_iterator
       &particle_two_way_coupling_force_cell,
     const typename DoFHandler<dim>::active_cell_iterator
-                                   &particle_velocity_cell,
+      &particle_velocity_cell,
+    const typename DoFHandler<dim>::active_cell_iterator
+                                   &particle_momentum_transfer_coefficient_cell,
     const VectorType               &particle_fluid_drag,
     const VectorType               &particle_fluid_force_two_way_coupling,
     const VectorType               &particle_velocity,
+    const VectorType               &particle_momentum_transfer_coefficient,
     const Parameters::DragCoupling &drag_coupling)
   {
     constexpr FEValuesExtractors::Vector vector_index(0);
@@ -1351,10 +1373,12 @@ public:
       .get_function_values(particle_fluid_force_two_way_coupling,
                            this->particle_two_way_coupling_force_values);
 
+    // particle_drag_values will remain zero in the implicit and semi-implicit
+    // coupling, since the momentum transfer coefficient is used instead, while
+    // particle_momentum_transfer_coefficient_values and
+    // particle_velocity_values will remain zero in the fully explicit coupling
     if (drag_coupling == Parameters::DragCoupling::fully_explicit)
-      { // particle_drag_values will remain zero in the implicit and
-        // semi-implicit coupling, since the momentum transfer
-        // coefficient is used instead
+      {
         this->fe_values_particle_drag->reinit(particle_drag_cell);
         (*this->fe_values_particle_drag)[vector_index].get_function_values(
           particle_fluid_drag, this->particle_drag_values);
@@ -1364,6 +1388,13 @@ public:
         this->fe_values_particle_velocity->reinit(particle_velocity_cell);
         (*this->fe_values_particle_velocity)[vector_index].get_function_values(
           particle_velocity, this->particle_velocity_values);
+
+        this->fe_values_particle_momentum_transfer_coefficient->reinit(
+          particle_momentum_transfer_coefficient_cell);
+        (*this->fe_values_particle_momentum_transfer_coefficient)
+          .get_function_values(
+            particle_momentum_transfer_coefficient,
+            this->particle_momentum_transfer_coefficient_values);
       }
   }
 
@@ -1511,12 +1542,15 @@ public:
   std::shared_ptr<FEValues<dim>> fe_values_particle_drag;
   std::shared_ptr<FEValues<dim>> fe_values_particle_two_way_coupling_force;
   std::shared_ptr<FEValues<dim>> fe_values_particle_velocity;
+  std::shared_ptr<FEValues<dim>>
+    fe_values_particle_momentum_transfer_coefficient;
 
   bool                        gather_particle_field_project;
   std::vector<Tensor<1, dim>> particle_velocity;
   std::vector<Tensor<1, dim>> particle_velocity_values;
   std::vector<Tensor<1, dim>> particle_drag_values;
   std::vector<Tensor<1, dim>> particle_two_way_coupling_force_values;
+  std::vector<double>         particle_momentum_transfer_coefficient_values;
   Tensor<1, dim>              average_particle_velocity;
   std::vector<Tensor<1, dim>> fluid_velocity_at_particle_location;
   std::vector<Tensor<1, dim>>